refactor(utils/throttler): max queue len to uint64

+++ node/throttled_compiler.go
@@ -15,7 +15,7 @@ type ThrottledCompiler struct {
 }
 
 func NewThrottledCompiler(
-	res compiler.Compiler, concurrencyBudget uint, maxQueueLen int32,
+	res compiler.Compiler, concurrencyBudget uint, maxQueueLen uint64,
 ) *ThrottledCompiler {
 	return &ThrottledCompiler{
 		Throttler: throttler.NewThrottler(
